refactor flow list to make it less insane
